Factorize JsonMessage into own file removing duplication
