+++ dotnet/dotNetGet.cs
@@ -15,7 +15,6 @@
 */
 
 using System;
-using Newtonsoft.Json;
 using IBM.XMS;
 
 
@@ -145,21 +144,6 @@ namespace ibmmq_samples
             destination.Dispose();
             sessionWMQ.Dispose();
             connectionWMQ.Close();
-        }
-
-        public class JsonMessage
-        {
-            public string msg;
-            public int random;
-            public JsonMessage(string s, int n)
-            {
-                msg = s;
-                random = n;
-            }
-            public string toJsonString()
-            {
-                return JsonConvert.SerializeObject(this);
-            }
         }
     }
 }
