+++ dotnet/dotNetSub.cs
@@ -15,7 +15,6 @@
 */
 
 using System;
-using Newtonsoft.Json;
 using IBM.XMS;
 
 namespace ibmmq_samples
@@ -144,21 +143,5 @@ namespace ibmmq_samples
             sessionWMQ.Dispose();
             connectionWMQ.Close();
         }
-
-        public class JsonMessage
-        {
-            public string msg;
-            public int random;
-            public JsonMessage(string s, int n)
-            {
-                msg = s;
-                random = n;
-            }
-            public string toJsonString()
-            {
-                return JsonConvert.SerializeObject(this);
-            }
-        }
-
     }
 }
