+++ dotnet/ibmmq_samples.csproj
@@ -72,6 +72,7 @@
     <Compile Include="dotNetResponse.cs" />
     <Compile Include="dotNetSub.cs" />
     <Compile Include="env.cs" />
+    <Compile Include="jsonMessage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Rocket.cs" />
   </ItemGroup>
